[EasySwoole] Fixed SymfonyServicesAppStateResetter.
