+++ packages/EasySwoole/src/Helpers/AppRuntimeHelper.php
@@ -27,14 +27,6 @@ final class AppRuntimeHelper
     }
 
     /**
-     * @param mixed[] $cacheTables
-     */
-    public static function setCacheTables(array $cacheTables): void
-    {
-        self::addOptions(['cache_tables' => $cacheTables]);
-    }
-
-    /**
      * @param callable[] $callbacks
      */
     public static function setCallbacks(array $callbacks): void
@@ -83,6 +75,11 @@ final class AppRuntimeHelper
         self::addOptions(['response_chunk_size' => $responseChunkSize]);
     }
 
+    public static function setRuntime(string $runtime): void
+    {
+        $_SERVER[self::APP_RUNTIME] = $runtime;
+    }
+
     /**
      * @param mixed[] $settings
      */
@@ -99,10 +96,5 @@ final class AppRuntimeHelper
     public static function setUseDefaultCallbacks(bool $useDefaultCallbacks): void
     {
         self::addOptions(['use_default_callbacks' => $useDefaultCallbacks]);
-    }
-
-    public static function setRuntime(string $runtime): void
-    {
-        $_SERVER[self::APP_RUNTIME] = $runtime;
     }
 }
